Back out previous, fix PR kern/49411 correctly.
frame->header.length is not initialized in the vendor packet case.
Instead, make hdmi_infoframe_header_pack take the packet length and
subtract the size of the header when packing the header.

(Could make it take the payload length instead, but for callers it is
more convenient to have a name for the whole packet length.)